Feat/porto tables

+++ src/apis/projects.ts
@@ -1,16 +1,10 @@
 import { createServerFn } from '@tanstack/react-start'
 import { getRequest } from '@tanstack/react-start/server'
-import { getServerAuth } from '#/infrastructure/auth'
+import { requireAdminAuth } from '#/infrastructure/auth'
 import { projectSchema } from '#/domain/schemas'
 import { listProjectsUseCase, createProjectUseCase, updateProjectUseCase, deleteProjectUseCase } from '#/domain/use-cases'
 import { drizzleProjectRepository } from '#/infrastructure/db/repositories/projectRepository'
 import { drizzleStatsRepository } from '#/infrastructure/db/repositories/statsRepository'
-
-async function requireUserId() {
-  const { userId } = await getServerAuth(getRequest())
-  if (!userId) throw new Error('Unauthorized')
-  return userId
-}
 
 export const ping = createServerFn({ method: 'GET' }).handler(async () => {
   console.log('[ping] handler called')
@@ -30,14 +24,14 @@ export const testDb = createServerFn({ method: 'GET' }).handler(async () => {
 })
 
 export const listProjects = createServerFn({ method: 'GET' }).handler(async () => {
-  await requireUserId()
+  await requireAdminAuth(getRequest())
   return listProjectsUseCase(drizzleProjectRepository)
 })
 
 export const createProject = createServerFn({ method: 'POST' })
   .validator((data: unknown) => projectSchema.parse(data))
   .handler(async ({ data }) => {
-    await requireUserId()
+    await requireAdminAuth(getRequest())
     return createProjectUseCase(drizzleProjectRepository, drizzleStatsRepository, data)
   })
 
@@ -47,13 +41,13 @@ export const updateProject = createServerFn({ method: 'POST' })
     return { id, data: projectSchema.partial().parse(rest) }
   })
   .handler(async ({ data }) => {
-    await requireUserId()
+    await requireAdminAuth(getRequest())
     return updateProjectUseCase(drizzleProjectRepository, data.id, data.data)
   })
 
 export const deleteProject = createServerFn({ method: 'POST' })
   .validator((data: unknown) => ({ id: String(data) }))
   .handler(async ({ data }) => {
-    await requireUserId()
+    await requireAdminAuth(getRequest())
     return deleteProjectUseCase(drizzleProjectRepository, drizzleStatsRepository, data.id)
   })
